Remove an misused rocksdb ticker (#2940)
The TickerType::NoIterators is actually used as a Guage in
RocksDB. When we require an iterator, the ticker increase 1, and when
we release the iterator, the ticker decrease 1. So if we reset the
ticker after we have required an iterator but before we have released
it, the ticker can be negative. Since we don't care about this ticker
yet, I just remove it here.