*: fix unknown file in log (#4714)
log 0.4 introduces non-static file path and module names. But it seems
to be a feature that most people don't use but break a lot of things.
This PR revert the lifetime back to static so that everything works
again.

Another approach is to update slog to support non-static module name and
file name. But it may lead to unnecessary clone and make things
complicated.

Know that this is just a workaround, we should switch back to crates io
version when slog is compatible with the latest log.

Signed-off-by: Jay Lee <busyjaylee@gmail.com>