copr: migrate cast decimal as time expr compatible with non-vec (#15185) (#8737)
Cast decimal to time may get wrong result because of different cast
policy with mysql. This commit fix it by add a parse_from_deciaml()
method to tidb_query_datatype::codec::mysql::Time, which deal with
decimal whole part as int, frac part as str.

The reason is that the parse logic of datetime is different when
dealing with strings and numbers. In short, when dealing with numbers,
if number length are not 6, 8, 12, 14, the parse logic should padded
with leading zeros to the closest length. But when dealing with strings,
the parse logic will interpreted from left to right to find year, month,
day, hour... for as many parts as are present in the string.

For more details, see
https://dev.mysql.com/doc/refman/5.7/en/date-and-time-literals.html.
Related tidb issue:
https://github.com/pingcap/tidb/issues/15185

Signed-off-by: lenshood <lenshood.zxh@gmail.com>

Co-authored-by: ti-srebot <66930949+ti-srebot@users.noreply.github.com>
Co-authored-by: Wenxuan <breezewish@pingcap.com>