raftstore: check valid leader before sleeping  (#9511)
*: check valid leader before sleeping

If leader is demoted and the region is hibernated, there can be no
leader for quite a long time. It's by design that client should try
other peer when leader is not available, but currently ticlient can't
handle the issue well. On the other hand, PD can still keep sending
scheduling command to the demoted leader, which will also timeout.

So this PR adds an additional check before sleeping to get around it.
This should fix the pd scheduling issue, but ticlient can still timeout
if demoted peer is restarted, client should still retry other peers when
necessary.

Close #9505.

Signed-off-by: Jay Lee <BusyJayLee@gmail.com>