raftstore: disable hibernate region except on master (#9735)
### What problem does this PR solve?

Problem Summary:

This PR adds a function to check if the binary is built on master
branch. Hibernate region is an experimental features that should
not be enabled on release branch, but it's enabled by accident on
5.0.0-rc. Checking branch name at build time can prevent such mistake
from happening again.

The side effect of the PR is binary built in PR will also have hibernate
regions disabled even it's checkout from master.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test
- Integration test

### Release note <!-- bugfixes or new feature need a release note -->
- disable hibernate region by default as it's still unstable