txn: make prewrite idempotent after a new lock is written (#12369)
close pingcap/tidb#34066

If a retried prewrite encounters a newer lock, the result may be not
idempotent.

This happens when the first prewrite succeeds and the transaction
is committed by resolving async-commit locks. Then, TiKV returns
KeyIsLocked in the retry while the first prewrite succeeds.

If the client does not handle it carefully, the client may report a false
commit failure to the user.

This commit fixes the bug, making prewrite idempotent in this case.

Signed-off-by: Yilin Chen <sticnarf@gmail.com>