engine_rocks: replace PerfLevel with PerfFlags in raftstore (#12373)
ref tikv/tikv#12362

PerfFlags provides more fine-grained PerfContext collecting. In this way, we
only collect the duration we really care about. Theoretically, it reduces the
overhead of collecting PerfContext compared to using PerfLevel::EnableTime.

This commit only changes the raftstore part. The read pool part will be done
later.

The default value of raftstore.perf-level is changed to Uninitialized. The
config is not documented before, so I think the change is not so important
to change the default value.

Signed-off-by: Yilin Chen <sticnarf@gmail.com>

Co-authored-by: Ti Chi Robot <ti-community-prow-bot@tidb.io>