*: fix multiple unit test errors in local environments (#12392)
close tikv/tikv#12390

1. Relax the IO load info assertion from "assert_eq" to assert "starts with".
So the test won't fail on Linux 4.19+ because of the extra fields.

2. Enable the `testexport` feature of `api_version` in the `dev-dependencies`
of `tikv`. Then, it is not necessary to specify the `testexport` feature manually
to compile tikv tests.

3.  For `plugin_registry` tests, search the dynamic library under the same
directory with the test binary. This avoids failures when people are running the
tests not under the cargo root or the default target directory path is changed.

Signed-off-by: Yilin Chen <sticnarf@gmail.com>

Co-authored-by: Ti Chi Robot <ti-community-prow-bot@tidb.io>