status_server: reload TLS certificate on TLS errors (#12569)
close tikv/tikv#12546

The commit implements auto TLS certificate reload for the status server.

The implementation is similar to ServerCredentialsFetcher of gRPC.
When there is a TLS error, we read the cert file metadata to check whether
the cert file has changed. If it changes, we re-initialize the SslContext
used for the HTTP server.

Signed-off-by: Yilin Chen <sticnarf@gmail.com>

Co-authored-by: Ti Chi Robot <ti-community-prow-bot@tidb.io>