raftstore: add metrics/logs to help debug high commit log duration (#13120)
ref tikv/tikv#13060, ref tikv/tikv#13078

In some cases, such as the one mentioned in #13078, the commit log duration 
became high. In the case, the needed log is not in entry cache and there are
many raftlog async fetch tasks.

This commit adds a log to show the cache first index and peers' progress when
there is any long uncommitted proposal. It also adds a metric to show the 
duration of the async fetch tasks.

Signed-off-by: cosven <yinshaowen241@gmail.com>

Co-authored-by: Ti Chi Robot <ti-community-prow-bot@tidb.io>