*: clean up tablet factory (#13912)
ref tikv/tikv#12842

In the past, there are 5 types of tablet factory:
- Dummy factory
- Test Factory for v1 and v2
- Production factory for v1 and v2.

Tablet factory also requires TabletAccessor trait, so there is 10 implementations.
If there are bugs, we need to change at least 5 places, and 10 places in worst
cases. This PR simplifies the code by limiting the scope of engine factory, so it
only needs to provide create, destroy and exist check three functions. All other
functions are moved to type `TabletRegistry`. Also `TabletFactory` is only for v2
usage, so there will be only 3 different tablet factories and no accessor trait.
It should be a lot easier to adjust the behavior of tablet factory.

Signed-off-by: Jay Lee <BusyJayLee@gmail.com>

Co-authored-by: Ti Chi Robot <ti-community-prow-bot@tidb.io>