compaction_guard: split SST when detected possible huge compaction (#15379)
close tikv/tikv#15058

This PR make the compaction guard splits SSTs when it find that there are possible huge compactions. It works by iterating the next of the output level (Let is be level L+1), when a SST crosses such a huge key range that making L+1 contains size greater than the `max-compaction-size`.

Signed-off-by: hillium <yujuncen@pingcap.com>

Co-authored-by: tonyxuqqi <tonyxuqi@outlook.com>