raftstore-v2: fix non-deterministic region merge (#15697)
close tikv/tikv#15682

This commit addresses the issue where a "region corrupted" error still
occurs in certain scenarios despite PR #15625 resolving the problem in
the transfer leader scenario. The root cause of the issue is the
non-deterministic nature of commit merge and rollback merge, allowing
transient errors during propose to trigger the problem again.

To fix this issue, the proposed solution ensures that TiKV only
initiates rollback merge when either the target region is not found or
the epoch has increased.

Signed-off-by: Neil Shen <overvenus@gmail.com>

Co-authored-by: tonyxuqqi <tonyxuqi@outlook.com>
Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>