test: fix unstable test test_limit_concurrency (#16361)
ref tikv/tikv#15990

The original test use select! to check t1 is finished before t2, but this is unstable when the select! is called when both t1 and t2 is finished. From the rust doc we can see " If multiple futures are ready, one will be pseudo-randomly selected at runtime."
which means if both future is ready, which is polled first is not determined. So this PR uses a atomic value to determine t1's complete to make the result stable.

Signed-off-by: glorv <glorvs@163.com>

Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>