tikv_util: make memory quota alloc and free simpler (#16938)
ref tikv/tikv#15990

Replace `compare_exchange_weak` loop with single `fetch_add` and `fetch_sub` to make the memory alloc and free simpler.
I do this change base on the fact that we does not expect the memory control to be precise on every operation. In high concurrency scenario and when the conflict is heavy, the real peak memory usage may be small than the capacity, but not smaller than `capacity - num_of_concurrency * avg_alloc_bytes`.

NOTE: In order to avoid counter overflow, this change introduces a hard limit(256PiB) that one alloc can acquire as it's big enough for any real world application. Passing higher value to `alloc` will result in an error, passing higher value to `alloc_force` and `free` will result in noop.

Signed-off-by: glorv <glorvs@163.com>