raftstore: add region_id to snap concurrency limiter APIs (#17051)
ref tikv/tikv#15972

An issue with the previous implementation of the concurrency limiter is that 
its APIs are not idempotent. If a single region leader keeps sending precheck 
requests to the same receiver (this may happen if the precheck response is 
somehow lost on the network), it would consume all reservations, thereby 
blocking all other snapshots. This commit addresses this problem by updating 
the concurrency limiter to deduplicate requests based on `region_id`. If the 
limiter already has a valid reservation for a region, calling try_recv() again 
will not allocate a new one.

Signed-off-by: Bisheng Huang <hbisheng@gmail.com>

Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>