raftstore: cancel gen snap task if a peer becomes unreachable (#17215)
close tikv/tikv#17191, close tikv/tikv#17213

After #17019, raft snapshots get blocked when a peer is removed
immediately after requesting a snapshot. This occurs because the leader
must wait for `MsgSnapGenPrecheckResponse` before generating snapshots.
As a result, not only is the snapshot blocked, but subsequent conf
changes (such as adding a node) are also blocked.

To resolve this issue, TiKV needs to cancel the snapshot generation
task if a peer becomes unreachable.

Signed-off-by: Neil Shen <overvenus@gmail.com>

Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>