raftstore: fix snap gen cancellation (#17231)
close tikv/tikv#17226

As a follow-up of #17215, this commit fixes an issue with the
`cancel_generating_snap` function to ensure that `snap_tried_cnt`
within PeerStorage is not unexpectedly incremented.

Background:
In PeerStorage, `snap_tried_cnt` is used to track the number of
snapshot generation attempts. As with what happened in #17226, TiKV
would panic if `snap_tried_cnt` exceeds the limit (currently hardcoded
to 5). By design, `snap_tried_cnt` should only be incremented when an
actual failure happens during the snapshot generation process. It
should not be incremented if the snapshot generation is intentionally
canceled.

Although its name suggests so, the `cancel_generating_snap` function
doesn't always cancel the snapshot generation. If the inputting
`compact_to` is so small that it doesn't make the snapshot stale,
`cancel_generating_snap` will return early and won't set `cancel` to
true. In the fix, we ensure that we only cancel the precheck when
`cancel` is set to true. Without it, `snapshot()` will see a 
`TryRecvError::Disconnected` error while `last_canceled` is false and 
increment `tried_cnt`.

Note that `cancel_generating_snap` is only called with a `compact_to`
value in `on_compact_raftlog`, which I believe to be the trigger of
issue #17226.

Also, this commit fixes a similar problem to ensure that `tried_cnt` is
not incremented in another potential scenario—when the peer requesting
snapshot is unknown.

Signed-off-by: Bisheng Huang <hbisheng@gmail.com>

Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>