In-memory Engine: Reduce `HybridEngine` scope (#17359)
In TiKV, the `KvEngine` trait provides the write path abstraction
(`WriteBatch`) and read path abstraction (`Snapshot`). The
`HybridEngine` implements the `KvEngine` trait, allowing it to be
integrated into TiKV.

However, as described in issue #17181, there is a major downside: all
components and features in TiKV use the `HybridEngine`, which can
introduce unexpected compatibility issues (e.g., BR restore failures
with checksum errors). This is not worthwhile because the `HybridEngine`
is designed specifically for accelerating coprocessor read speed in
scenarios with many MVCC workloads.

This PR shrinks the `HybridEngine` scope from being globally enabled to
only being enabled in the coprocessor. The changes include:
- Adding a new observer, `WriteBatchObserver`, to observe write requests
  via raftstore coprocessor hooks.
- Adding a new observer, `SnapshotObserver`, to observe snapshot (and
  `read_ts`) via raftstore coprocessor hooks.
- Adding a new method to the `RaftKv` struct called `async_in_memory_snapshot`,
  which takes snapshots from the `HybridEngine`.

This PR also introduces an architectural change, transforming the
in-memory engine from functioning as an `KvEngine` to acting as an
MVCC-aware cache. Conceptually, the in-memory engine is shifted from
the storage engine layer to the MVCC layer. This adjustment better
aligns with its purpose of improving MVCC performance by caching recent
versions, rather than serving as a full-fledged storage engine.

    ┌──────────────┐      ┌──────────────┐
    │ Copr/Storage │      │ Copr/Storage │
    ├──────────────┤      ├──────────────┤
    │    MVCC      │      │     MVCC     │
    ├──────────────┤      │ ┌──────────┐ │
    │ Raftstore    │ ───> │ │   IME    │ │
    ├──────────────┤      ├─┴──────────┴─┤
    │    IME       │      │  Raftstore   │
    │┌────────────┐│      ├──────────────┤
    ││  RocksDB   ││      │    RocksDB   │
    └┴────────────┴┘      └──────────────┘

Additionally, this PR brings two unexpected benefits*:
- Reduced TiKV binary size: From 1.7GB to 1.2GB.
- Faster build time: From 32 minutes to 20 minutes.

*: Results were obtained using the `make release` command.

Signed-off-by: Neil Shen <overvenus@gmail.com>