raftstore: fix panic due to stale peer handling snapshot (#17535)
close tikv/tikv#17469

The commit fixes a panic in TiKV that occurs in a rare scenario that
involves region splits and immediate removal of the new peer.

When a region splits, the new peer on a follower can be created in two 
ways: (1) By receiving a Raft message from the new region 
(`fn maybe_create_peer`) (2) By applying the split operation locally 
(`fn on_ready_split_region`). 

Depending on timing, a new peer might first be created by a Raft 
message and then again when the split is applied. This is a known 
situation. When it happens, the second peer replaces the first, and the 
first peer is dicarded. However, the discarded peer may continue 
processing existing messages, leading to unexpected states.

The panic can be reproduced with the following sequence of events:
1. The first peer is created by a Raft message and is waiting for a 
   Raft snapshot.
2. The second peer (of the same reason) is created by 
   `on_ready_split_region` when the split operation is applied, 
   replacing the first peer and closing its mailbox (as expected).
3. The second peer is immediately removed (a rare situation). This 
   removes the region metadata.
4. The first peer continues processing the Raft snapshot message, 
   expecting the metadata of the region to exist, causing the panic.

Signed-off-by: Bisheng Huang <hbisheng@gmail.com>

Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>