In-memory Engine: fix panic that new region has the same id with evicting region (#17550)
close tikv/tikv#17487, close tikv/tikv#17493, close tikv/tikv#17548

Fix the bug that in-memory-engine may panic when load new region whose id is the same with a cached region that is being evicted. This can happen when a region is evicted by region merge, and the merged region is split and then load again while the cached region is still in evicting state.

Signed-off-by: glorv <glorvs@163.com>