grafana: clarify raft waterfall metrics descriptions (#17766)
ref tikv/tikv#15990

Raft waterfall metrics track the duration of individual requests, all 
beginning from the same starting point (when the async write request is 
scheduled) but ending at various stages of the write process. Previous 
descriptions did not make that clear and may confuse the readers. This 
commit improves the grafana descriptions for clarity.

Signed-off-by: Bisheng Huang <hbisheng@gmail.com>

Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>