raftstore: do not override min_matched index with min_committed index (#18204)
close tikv/tikv#18129

During region merge, the CacheUpLogs phase is used to ensure all peers can replicate all the raft logs even if leader is already merged. The original low bound of these logs is calculated by the min matched index of all peers. But in #9982, the bound is replaced with the min committed index, but this is not correct because committed index can not ensure raft state can be recovered from the CatchUpLogs message when the current leader is restarted after proposing PrepareMerge.

Signed-off-by: glorv <glorvs@163.com>

Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>