txn: return debug info for some error when `commit_role` is unknown (#18448)
close tikv/tikv#18441

In the previous PR, we only collect mvcc info when `commit_role` is `Secondary` when `commit_ts < min_commit_ts`. However, when resolving a lock, the `commit_role` of `commit` is None and we cannot get any mvcc info when this error happens. This PR does an enhancement and checks whether the resolved key is primary in the lock, if not, it still collect mvcc for further debugging.

Signed-off-by: Chao Wang <cclcwangchao@hotmail.com>
Signed-off-by: 王超 <cclcwangchao@hotmail.com>

Co-authored-by: cfzjywxk <cfzjywxk@gmail.com>