titan: Estimate raw blob size based on Titan metrics (#18628)
close tikv/tikv#18600

This PR introduces a mechanism to estimate Titan blob raw size based on
metrics collected from Titan during flush. 

A sliding window of recent (raw, compressed) size samples is used to 
compute an average expansion factor. This factor is then used to 
estimate the raw size of a blob when building SST table properties. 
This approach assumes that the compression ratio remains relatively 
stable for the workload.

Apparently this is not a perfect solution, but a workaround. A more 
direct solution—recording the uncompressed size in the blob index 
format—was previously proposed[1]. However, changing the data format 
would break backward compatibility and prevent rollbacks, which can be 
risky.

[1] https://github.com/tikv/titan/pull/334#issuecomment-3026813148

Signed-off-by: Bisheng Huang <hbisheng@gmail.com>