config: adjust default value of config `rocksdb.max-manifest-file-size`  from 128MiB to 256MiB (#18496)
ref tikv/tikv#15990

The default value of config `rocksdb.max-manifest-file-size` is too small for large clusters with more than 100k sst files. It can cause long ingest latency when the effective manifest file size exceed the threshold. So this PR adjust the default value to 256MiB which should be a large enough value.

Signed-off-by: glorv <glorvs@163.com>